backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.2.0


celery==5.3.4
redis==5.0.1


django-celery-results==2.5.1




pytest==7.4.2
pytest-django==4.5.2

PyJWT==2.10.1